rename to cartesian animation
